[IMP] web: convert `view_button` tests to HOOT
task-3705027
